requirements.txt
anyio==3.6.1
appdirs==1.4.4
asgiref==3.4.1
attrs==21.4.0
backports.cached-property==1.0.2
black==19.10b0
btcpay-python==1.3.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Django==3.2.5
django-admin-display==1.3.0
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-picklefield==3.1
ecdsa==0.18.0
graphql-core==3.2.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.7.0
Pillow==9.2.0
pip==22.1.2
psycopg2-binary==2.9.3
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
regex==2022.7.9
requests==2.28.1
setuptools==61.2.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
starlette==0.20.4
strawberry-django-jwt==0.2.2
strawberry-graphql==0.123.2
strawberry-graphql-django==0.4.0
toml==0.10.2
typed-ast==1.4.3
typing_extensions==4.1.1
urllib3==1.26.12
uvicorn==0.18.2
wheel==0.37.1
